requirements.txt
beautifulsoup4==4.5.1
PyYAML==3.12
requests==2.11.1
APScheduler==3.2.0